+++ components/forms/FilterableSearch/react.test.js
@@ -14,7 +14,7 @@ describe('FilterableSearch', () => {
     props = {
       onFilterChange: jest.fn(),
       onInputChange: jest.fn(),
-      onSearch: jest.fn(),
+      onSubmit: jest.fn(),
       value: '',
       filterChoices: [
         { label: 'Filter 1', value: 'FILTER1', checked: false },
@@ -41,20 +41,13 @@ describe('FilterableSearch', () => {
       wrapper.find('DropdownToggle').simulate('click')
       expect(wrapper.find('DropdownMenu').prop('show')).toEqual(false)
     })
-    it('should hide the filter dropdown when clicking outside the menu', () => {
-      expect(wrapper.find('DropdownMenu').prop('show')).toEqual(false)
-      wrapper.find('DropdownToggle').simulate('click')
-      expect(wrapper.find('DropdownMenu').prop('show')).toEqual(true)
-      wrapper.find('Input').simulate('click') // works in browser, but this simulated click doesn't seem to work
-      expect(wrapper.find('DropdownMenu').prop('show')).toEqual(false)
-    })
     it('should trigger search action when the user hits enter in input or clicks SearchButton', () => {
-      expect(props.onSearch).toBeCalledTimes(0)
+      expect(props.onSubmit).toBeCalledTimes(0)
       wrapper.find('SearchButton').simulate('click')
-      expect(props.onSearch).toBeCalledTimes(1)
-      wrapper.find('.input').simulate('keyDown', { keyCode: 1 }) // not Enter key
-      wrapper.find('.input').simulate('keyDown', { keyCode: 13 }) // Enter key
-      expect(props.onSearch).toBeCalledTimes(2)
+      expect(props.onSubmit).toBeCalledTimes(1)
+      wrapper.find('.text-field__input').simulate('keyDown', { keyCode: 1 }) // not Enter key
+      wrapper.find('.text-field__input').simulate('keyDown', { keyCode: 13 }) // Enter key
+      expect(props.onSubmit).toBeCalledTimes(2)
     })
   })
 })