+++ packages/external-ui-react/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 2019-03-01
+
+### Added
+
+- New label and placeholder functionality and design to `TextField`.
+
+### Changed
+
+- Rename `Input` to `TextField`. `Input` still works.
+- `TextField` error state icon from an X to an alert triangle.
+
 ## [1.6.0] - 2019-02-22
 
 ### Changed